Add files via upload
This generator reproduces the procedure for generating random encounters found in the Dolmenwood Campaign Book 2022-11-28, from Gavin Norman/Necrotic Gnome).

The Dolmenwood setting is still a work in progress and is going to become a standalone rule-system, so this generator applies only to those using the Old-School Essentials PDF versions of the setting.

It's a very simple project. Future goals for it will be making it calculate the number of creatures appearing, rather than requiring the Referee to roll dice to determine. It's a simple upgrade but I felt like just putting out something that work and updating later.